add material components
