+++ package.json
@@ -15,7 +15,9 @@
     "react": "17.0.2",
     "react-native": "0.66.4",
     "react-native-bouncy-checkbox": "^2.1.10",
-    "react-native-uuid": "^2.0.1"
+    "react-native-paper": "^4.11.2",
+    "react-native-uuid": "^2.0.1",
+    "react-native-vector-icons": "^9.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
